+++ tests/test-phoenix_respiratory.R
@@ -23,18 +23,12 @@ stopifnot(identical(phoenix_respiratory(data = sepsis), 0L))
 ################################################################################
 # verify error if lengths differ
 x <- tryCatch(phoenix_respiratory(pf_ratio = numeric(0)), error = function(e) e)
-stopifnot(inherits(x, "simpleError"))
-stopifnot(identical(
-  x$message,
- "All inputs need to either have the same length or have length 1. Length of pf_ratio is 0; Length of sf_ratio is 1; Length of imv is 1; Length of other_respiratory_support is 1."
-))
+stopifnot(isTRUE(inherits(x, "error")))
+stopifnot(grepl("All inputs need to either have the same length or have length 1.", x$message))
 
 x <- tryCatch(phoenix_respiratory(pf_ratio = c(NA, NA), imv = c(NA, NA, NA)), error = function(e) e)
-stopifnot(inherits(x, "simpleError"))
-stopifnot(identical(
-  x$message,
- "All inputs need to either have the same length or have length 1. Length of pf_ratio is 2; Length of sf_ratio is 1; Length of imv is 3; Length of other_respiratory_support is 1."
-))
+stopifnot(isTRUE(inherits(x, "error")))
+stopifnot(grepl("All inputs need to either have the same length or have length 1.", x$message))
 
 ################################################################################
 #                                 End of File                                  #